Create zomato data analysis and graphical visulaisation
The goal of this project is to perform data analysis on a dataset containing information about various restaurants in Banglore . The dataset includes several columns such as 'name,' 'rate,' 'cuisine,' etc. The primary objective is to gain insights and visualize key aspects of the restaurant data.
Tools Used:
Python
Pandas for data manipulation
Matplotlib and Seaborn for data visualization
Conclusion:
This exploratory data analysis project provides valuable insights into various aspects of the restaurant data, allowing for a better understanding of the distribution of ratings, preferences in different locations, and the relationship between different variables.